chore: refactor daemonsets
