--- frontend/src/container/InfraMonitoringK8s/DaemonSets/DaemonSetDetails/Events/index.ts
@@ -1,3 +1,0 @@
-import DaemonSetEvents from './DaemonSetEvents';
-
-export default DaemonSetEvents;
