--- frontend/src/container/InfraMonitoringK8s/DaemonSets/DaemonSetDetails/Logs/index.ts
@@ -1,3 +1,0 @@
-import DaemonSetLogs from './DaemonSetLogsDetailedView';
-
-export default DaemonSetLogs;
