--- frontend/src/container/InfraMonitoringK8s/DaemonSets/DaemonSetDetails/Metrics/index.ts
@@ -1,3 +1,0 @@
-import DaemonSetMetrics from './DaemonSetMetrics';
-
-export default DaemonSetMetrics;
