--- frontend/src/container/InfraMonitoringK8s/DaemonSets/DaemonSetDetails/Traces/index.ts
@@ -1,3 +1,0 @@
-import DaemonSetTraces from './DaemonSetTraces';
-
-export default DaemonSetTraces;
