+++ frontend/src/container/InfraMonitoringK8s/EntityDetailsUtils/utils.tsx
@@ -25,6 +25,7 @@ export const QUERY_KEYS = {
 	K8S_DEPLOYMENT_NAME: 'k8s.deployment.name',
 	K8S_STATEFUL_SET_NAME: 'k8s.statefulset.name',
 	K8S_JOB_NAME: 'k8s.job.name',
+	K8S_DAEMON_SET_NAME: 'k8s.daemonset.name',
 };
 
 /**
